+++ mobile/android/android-components/components/feature/media/src/test/java/mozilla/components/feature/media/fullscreen/MediaSessionFullscreenFeatureTest.kt
@@ -9,17 +9,22 @@ import android.content.pm.ActivityInfo
 import android.os.Build
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import mozilla.components.browser.state.action.ContentAction
+import mozilla.components.browser.state.action.CustomTabListAction
 import mozilla.components.browser.state.action.MediaSessionAction
 import mozilla.components.browser.state.action.TabListAction
 import mozilla.components.browser.state.state.BrowserState
 import mozilla.components.browser.state.state.MediaSessionState
+import mozilla.components.browser.state.state.SessionState
+import mozilla.components.browser.state.state.createCustomTab
 import mozilla.components.browser.state.state.createTab
 import mozilla.components.browser.state.store.BrowserStore
 import mozilla.components.concept.engine.mediasession.MediaSession
+import mozilla.components.support.test.ext.joinBlocking
 import mozilla.components.support.test.libstate.ext.waitUntilIdle
 import mozilla.components.support.test.mock
 import mozilla.components.support.test.rule.MainCoroutineRule
 import org.junit.Assert.assertEquals
+import org.junit.Assert.assertNotEquals
 import org.junit.Assert.assertTrue
 import org.junit.Rule
 import org.junit.Test
@@ -55,7 +60,8 @@ class MediaSessionFullscreenFeatureTest {
         val store = BrowserStore(initialState)
         val feature = MediaSessionFullscreenFeature(
             activity,
-            store
+            store,
+            null
         )
 
         feature.start()
@@ -84,7 +90,8 @@ class MediaSessionFullscreenFeatureTest {
         val store = BrowserStore(initialState)
         val feature = MediaSessionFullscreenFeature(
             activity,
-            store
+            store,
+            null
         )
 
         feature.start()
@@ -113,7 +120,8 @@ class MediaSessionFullscreenFeatureTest {
         val store = BrowserStore(initialState)
         val feature = MediaSessionFullscreenFeature(
             activity,
-            store
+            store,
+            null
         )
 
         feature.start()
@@ -144,7 +152,8 @@ class MediaSessionFullscreenFeatureTest {
         val store = BrowserStore(initialState)
         val feature = MediaSessionFullscreenFeature(
             activity,
-            store
+            store,
+            null
         )
 
         feature.start()
@@ -181,7 +190,8 @@ class MediaSessionFullscreenFeatureTest {
         val store = BrowserStore(initialState)
         val feature = MediaSessionFullscreenFeature(
             activity,
-            store
+            store,
+            null
         )
 
         feature.start()
@@ -230,7 +240,8 @@ class MediaSessionFullscreenFeatureTest {
         val store = BrowserStore(initialState)
         val feature = MediaSessionFullscreenFeature(
             activity,
-            store
+            store,
+            null
         )
 
         feature.start()
@@ -253,5 +264,60 @@ class MediaSessionFullscreenFeatureTest {
         store.waitUntilIdle()
 
         assertEquals(ActivityInfo.SCREEN_ORIENTATION_USER, activity.requestedOrientation)
+    }
+
+    @Suppress("Deprecation")
+    @Test
+    @Config(sdk = [Build.VERSION_CODES.N])
+    fun `GIVEN the currently selected tab is in pip mode WHEN a custom tab loads THEN display custom tab in device's current orientation`() {
+        val activity = Robolectric.buildActivity(Activity::class.java).setup().get()
+        val elementMetadata = MediaSession.ElementMetadata()
+        val initialState = BrowserState(
+            tabs = listOf(
+                createTab(
+                    "https://www.mozilla.org", id = "tab1",
+                    mediaSessionState = MediaSessionState(
+                        mock(),
+                        elementMetadata = elementMetadata,
+                        playbackState = MediaSession.PlaybackState.PLAYING,
+                        fullscreen = true
+                    )
+                )
+            ),
+            selectedTabId = "tab1"
+        )
+        val store = BrowserStore(initialState)
+
+        val feature = MediaSessionFullscreenFeature(
+            activity,
+            store,
+            null
+        )
+
+        feature.start()
+        activity.enterPictureInPictureMode()
+        store.waitUntilIdle()
+
+        store.dispatch(ContentAction.PictureInPictureChangedAction("tab1", true))
+        store.waitUntilIdle()
+        assertEquals(ActivityInfo.SCREEN_ORIENTATION_UNSPECIFIED, activity.requestedOrientation)
+
+        val customTab = createCustomTab(
+            "https://www.mozilla.org",
+            source = SessionState.Source.Internal.CustomTab,
+            id = "tab2"
+        )
+        store.dispatch(CustomTabListAction.AddCustomTabAction(customTab)).joinBlocking()
+        val externalActivity = Robolectric.buildActivity(Activity::class.java).setup().get()
+        assertEquals(1, store.state.customTabs.size)
+        store.waitUntilIdle()
+        val featureForExternalAppBrowser = MediaSessionFullscreenFeature(
+            externalActivity,
+            store,
+            "tab2"
+        )
+        featureForExternalAppBrowser.start()
+
+        assertNotEquals(ActivityInfo.SCREEN_ORIENTATION_LANDSCAPE, externalActivity.requestedOrientation)
     }
 }